creates helper function for populate listings with data
